chore: refactor ad
